% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ncdf_output.R
\name{add_netcdf_output}
\alias{add_netcdf_output}
\title{Add output from model-specific model runs to  netcdf file}
\usage{
add_netcdf_output(output_lists, folder = ".", model, out_file)
}
\arguments{
\item{output_lists}{list; list containing lists of 
output (e.g. temperature, ice_height)}

\item{folder}{filepath; to folder which contains the model folders
generated by export_config()}

\item{model}{vector; model to export driving data. Options 
include c("GOTM", "GLM", "Simstrat",
"FLake", "MyLake")}

\item{out_file}{filepath; to save netCDF file defaults to "ensemble_output.nc"}
}
\description{
Add data from lists of output, generated by run_ensemble to  a
netcdf file created by earlier
run of run_ensemble
}
\keyword{internal}
