% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ncdf.R
\name{analyse_ncdf}
\alias{analyse_ncdf}
\title{Analyse the output netCDF}
\usage{
analyse_ncdf(ncdf, spin_up = NULL)
}
\arguments{
\item{ncdf}{filepath; to the `ensemble_output.nc` file}

\item{spin_up}{numeric; Number of days to disregard as spin-up for analysis.}
}
\value{
list with the dataframes used for comparison, statistics on model performance,
stratification
}
\description{
Analyse the LER output neCDF and produce summary statistics
}
\examples{
\dontrun{
}
}
