% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmix.R
\encoding{UTF-8}
\name{calc_hmix}
\alias{calc_hmix}
\title{Estimate mixed-layer depth from temperature profiles}
\usage{
calc_hmix(
  temps,
  depths,
  min.dT = 1,
  therm.dz = 1,
  min.hmix = 1.5,
  plot = FALSE,
  ...
)
}
\arguments{
\item{temps}{a numeric vector of water temperature in degC.}

\item{depths}{a numeric vector corresponding to the depths (in m) of the temps
measurements}

\item{min.dT}{numeric; minimum change in water temperature}

\item{therm.dz}{numeric; depth resolution for calculating thermoclime}

\item{min.hmix}{numeric; minimum depth of hmix otherwise returns NA. Defaults to 1.5m}

\item{plot}{boolean; plot temperature profile with calculated mixed depth.
Defaults to FALSE}

\item{...}{arguments to be passed to base plot}
}
\description{
This function estimates the mixed depth. It finds the minimum curvature
in the profile (maximum for winter stratification) as the border of the surface mixed
layer. Then it finds the thermocline as the depth of maximum T-gradient. It 
estimates the mixed layer depth as the depth where the regression line through
the surface layer temperatures intersects with the regression line through the
thermocline temperatures. It performs some other checks, like whether 
stratification exists (Ts-Tb) > threshold min.dT, and whether the thermocline 
is above the surface layer, or whether the lake is completely isothermal and
there is no intersection (and returns NA). If mixed, it assumes the mixed 
layer depth is the maximum depth. It also plots the profile if desired 
(plot=TRUE), marking the surface layer and inflection point in red and the 
thermocline in blue.
}
\author{
Tom Shatwell
}
