% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali_ensemble.R
\name{change_pars}
\alias{change_pars}
\title{Change parameter or meteo scaling for a model}
\usage{
change_pars(config_file, model, pars, type, met, folder)
}
\description{
Input a specific parameter or scale the meteorological forcing for a selected model
}
