% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_output.R
\name{connect_output}
\alias{connect_output}
\title{Connect model output}
\usage{
connect_output(model, config_file, folder = ".")
}
\arguments{
\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{config_file}{name of the master LakeEnsemblR config file}

\item{folder}{folder}
}
\value{
list
}
\description{
Read in FLake results, interpolate to a gridded struture and return a dataframe in long or wide format for vars = "temp"
}
\examples{
\dontrun{
}

}
