% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calibration.R
\name{cost_model}
\alias{cost_model}
\title{Run a model and calculate model cost}
\usage{
cost_model(
  config_file,
  model,
  var,
  folder,
  obs_deps,
  obs_out,
  out_hour,
  qualfun,
  config_f
)
}
\arguments{
\item{config_file}{path to master config file}

\item{model}{name of the model}

\item{var}{name of variable for which to calculate model performance}

\item{folder}{root folder}

\item{obs_deps}{depths of observations}

\item{obs_out}{data.frame with the observations of `var`}

\item{out_hour}{FLake specific: hout of output}

\item{qualfun}{function that takes data.frames of observations and simulations and returns
model performance metrics}

\item{config_f}{path to model specific control file (e.g. "GLM/glm3.nml")}
}
\description{
Runns the selected model and calculates fit metrics using a provided funtion
}
\keyword{internal}
