% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_netcdf_output.R
\name{create_netcdf_output}
\alias{create_netcdf_output}
\title{Create netcdf output from model-specific model runs}
\usage{
create_netcdf_output(
  output_lists,
  folder = ".",
  out_time,
  longitude = 0,
  latitude = 0,
  compression = 4,
  out_file = "ensemble_output.nc"
)
}
\arguments{
\item{output_lists}{list; list containing lists of output (e.g. temperature, ice_height)}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{out_time}{data frame; data frame with column ("datetime"),
describing at what times output should be generated}
}
\description{
Create a netcdf from lists of output, generated by run_ensemble
}
