% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_extinction.R
\name{export_extinction}
\alias{export_extinction}
\title{Export extinction coefficients}
\usage{
export_extinction(config_file, model = c("GOTM", "GLM", "Simstrat", "FLake"))
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}
}
\description{
Exports extinction coefficients for each model based on a master LakeEnsemblR config file
}
\keyword{methods}
