% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_flow.R
\name{export_flow}
\alias{export_flow}
\title{Export LakeEnsemblR standardised flow files to model specific driver format}
\usage{
export_flow(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = "."
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include
c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}
}
\description{
Export in- anbd out-flow driver files for each model
}
\examples{
\dontrun{
export_flow(config_file, model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"))
}
}
