% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_init_cond.R
\name{export_init_cond}
\alias{export_init_cond}
\title{Export initial conditions into each model setup from the LakeEnsemblR standardized input for observed temperature profile}
\usage{
export_init_cond(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  date = NULL,
  print = TRUE
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{date}{character; Date in "YYYY-mm-dd HH:MM:SS" format to extract the initial profile.
If NULL, the observations file specified in config_file is used to extract
the date.}

\item{print}{logical; Prints the temperature profile to the console}
}
\description{
Export initial condition files and input into model configuration files.
}
