% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_location.R
\name{export_location}
\alias{export_location}
\title{Export location settings for each model}
\usage{
export_location(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = "."
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{filepath; to folder which contains the model folders generated
by export_config()}
}
\description{
Exports settings like hypsograph, coordinates and ice settings,
 based on the master LakeEnsemblR config file
}
\examples{

}
\keyword{methods}
