% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_meteo.R
\name{export_meteo}
\alias{export_meteo}
\title{Export LakeEnsemblR standardized input to model specific driver format}
\usage{
export_meteo(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}
}
\description{
Export driver files for each model
}
\examples{
\dontrun{
export_meteo(model = c('GOTM', 'GLM', 'Simstrat', 'FLake'),
             meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
}
