% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_output_settings.R
\name{export_output_settings}
\alias{export_output_settings}
\title{Export output settings for each model}
\usage{
export_output_settings(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = "."
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{filepath; to folder which contains the model folders generated
by export_config()}
}
\description{
Exports settings related to output (time step, format),
 for each model
}
\examples{

}
\keyword{methods}
