% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_time.R
\name{export_time}
\alias{export_time}
\title{Export time settings for each model}
\usage{
export_time(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{folder}
}
\description{
Exports settings like start and end time and time step,
 based on the master LakeEnsemblR config file
}
\keyword{methods}
