% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_value.R
\name{get_config_value}
\alias{get_config_value}
\title{Extract value from any LER model config file}
\usage{
get_config_value(model, file, label, key)
}
\arguments{
\item{model}{string; name of a model in the LER package}

\item{file}{filepath; to model-specific config file}

\item{label}{string; which corresponds to section where the key is located}

\item{key}{string; name of key in which to extract the value}
}
\description{
Wrapper function for functions that extract values from model config files
}
\examples{
\dontrun{
get_config_value(model = "GOTM", file = "gotm.yaml", label = "turb_param", key = "k_min")
}

}
