% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json_value.R
\name{get_json_value}
\alias{get_json_value}
\title{Extracts values from a json file}
\usage{
get_json_value(file, label, key, out_file = NULL)
}
\arguments{
\item{file}{filepath; to json file}

\item{label}{string; which corresponds to section where the key is located}

\item{key}{string; name of key in which to extract the value}
}
\description{
Extracts values from a json file by locating the label and key within the json file.
}
\examples{
\dontrun{
input_json(file = "samp.par", label = "ModelParameters", key = "f_wind", value = 1.2, out_file = NULL)
}
}
\author{
Tadhg Moore, Jorrit Mesman
}
