% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calibration.R
\name{get_model_met_name}
\alias{get_model_met_name}
\title{get the name of model meteo file}
\usage{
get_model_met_name(model, config_f)
}
\arguments{
\item{model}{Model name}

\item{config_f}{path to model specific controll file}
}
\description{
get the name of the meteo file from the model config file
}
\keyword{internal}
