% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mylakeconfig_value.R
\name{get_mylakeconfig_value}
\alias{get_mylakeconfig_value}
\title{Extracts values from the MyLake config file}
\usage{
get_mylakeconfig_value(file, label, key)
}
\arguments{
\item{file}{filepath; to R object (loaded Rdata file)}

\item{label}{string; which corresponds to section where the key is located}

\item{key}{string; name of key in which to extract the value}
}
\description{
Extracts values from the MyLake config file by locating the label and key within the file.
}
