% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Get output data for each model}
\usage{
get_output(config_file, model, vars, obs_depths = NULL, folder = ".")
}
\arguments{
\item{config_file}{filepath; To LER config yaml file. Only used if model = 'GOTM'}

\item{model}{character; Model for which scaling parameters will be applied. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{vars}{vector; variables to extract from FLake output. Currently just temp and ice}

\item{obs_depths}{vector; Observation depths. Defaults to NULL}

\item{folder}{filepath; to folder which contains the model folders generated by export_config(). Defaults to '.'}
}
\value{
dataframe or list of output variables
}
\description{
Get output data for each model
}
