% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simstrat_temp.R
\name{get_simstrat_temp}
\alias{get_simstrat_temp}
\title{Get temperature data for Simstrat}
\usage{
get_simstrat_temp(folder = ".", par_file = "simstrat.par", obs_depths = NULL)
}
\arguments{
\item{folder}{filepath; to folder which contains the model folders generated by export_config().
Defaults to '.'}

\item{par_file}{filename; of simstrat JSON file. Defaults to 'simstrat.par'}

\item{obs_depths}{vector; Observation depths to include in output. Defaults to NULL}
}
\value{
dataframe in rLakeAnalyzer format.
}
\description{
Get temperature data for Simstrat
}
