% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wtemp_df.R
\name{get_wtemp_df}
\alias{get_wtemp_df}
\title{Extract FLake output as a dataframe}
\usage{
get_wtemp_df(output, depths, folder = ".", nml_file, long = FALSE)
}
\arguments{
\item{output}{filepath; to FLake output file}

\item{depths}{vector; of numeric values to extract the depths at}

\item{nml_file}{filepath; to FLake namelist file}

\item{long;}{Boolean; return data in long form}
}
\value{
Dataframe in rLakeAnalyzer format
}
\description{
Read in FLake results, interpolate to a gridded struture and return a dataframe in the format of rLakeAnalyzer
}
\examples{
\dontrun{
df <- get_wtemp_df(output = 'FLake/output/output.dat', folder = 'FLake', nml_file = 'FLake/feeagh.nml')
}

}
