% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yaml_value.R
\name{get_yaml_value}
\alias{get_yaml_value}
\title{Get a value from the loaded yaml file}
\usage{
get_yaml_value(yaml, ...)
}
\arguments{
\item{yaml}{list; loaded using `read_yaml()`}

\item{...}{character string with the keys from the yaml file. They need to be in consecutive order}
}
\value{
list with updated yaml parameters
}
\description{
Get a value from a loaded yaml file.
}
\examples{

\dontrun{
config_file <- system.file("extdata/feeagh/LakeEnsemblR.yaml", package = "LakeEnsemblR")

yaml <- read_yaml(config_file)

lat <- get_yaml_value(yaml, "location", "latitude")
kmin_init <- get_yaml_value(yaml, "calibration", "GOTM", "turb_param/k_min", "initial")
}
}
\author{
Tadhg Moore
}
