% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_json.R
\name{input_json}
\alias{input_json}
\title{Input values into json file}
\usage{
input_json(file, label, key, value, arg_list = NULL, out_file = NULL)
}
\arguments{
\item{file}{filepath; to json file which you wish to edit}

\item{label}{string; which corresponds to section where the key is located}

\item{key}{string; name of key in which to input the value}

\item{value}{string; to be input into the key/value pair.
Note boolean values must be input as 'true'/'false' as per the json format}

\item{arg_list}{list; of parameter to be updated. List names need to be the same
as the key (e.g. 'a_seiche'.}

\item{out_file}{filepath; to write the output json file (optional);
defaults to overwriting file if not specified}
}
\description{
Inputs values into json file by locating the label and key within the json file.
 Preserves comments (!) if present.
 NOTE: this does not use a json parser so if there are json formatting errors
 this function will not pick them up.
}
\examples{

\dontrun{
input_json(file = "samp.par", label = "ModelParameters", key = "f_wind",
 value = 1.2, out_file = NULL)
}
}
\author{
Tadhg Moore, Jorrit Mesman
}
