% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_mylakeconfig.R
\name{input_mylakeconfig}
\alias{input_mylakeconfig}
\title{Inputs value into the MyLake config file}
\usage{
input_mylakeconfig(file, label, key, value, out_file = NULL)
}
\arguments{
\item{file}{filepath; to R object (loaded Rdata file)}

\item{label}{string; which corresponds to section where the key is located}

\item{key}{string; name of key in which to extract the value}

\item{value}{string; name of key in which to extract the value}

\item{out_file}{filepath; to write the output config file (optional); defaults to overwriting file if not specified}
}
\description{
Inputs value into the MyLake config file by locating the label and key within the file.
}
