% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{inspect_output}
\alias{inspect_output}
\title{Inspect model output within a Shiny application}
\usage{
inspect_output(model, config_file)
}
\arguments{
\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{config_file}{name of the master LakeEnsemblR config file}
}
\value{
Launches a shiny application
}
\description{
Loads raw model output and allows for inspection withing a Shiny application.
}
\examples{
\dontrun{
}

}
