% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_var_dic.R
\docType{data}
\name{lake_var_dic}
\alias{lake_var_dic}
\title{Lake variable dictionary}
\format{
## `lake_var_dic`
A data frame with 8 rows and 3 columns:
\describe{
  \item{standard_name}{Name used for each variable}
  \item{short_name}{Short name used within the code and netCDF file}
  \item{unit}{Unit used for each variable}
}
}
\source{
Developer generated.
}
\usage{
lake_var_dic
}
\description{
Dictionary for lake variables.
}
\keyword{datasets}
