% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LHC.R
\name{load_LHC_results}
\alias{load_LHC_results}
\title{Plot results of LHC calibration}
\usage{
load_LHC_results(config_file, model, res_files)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file for which `cali_ensemble()`
was run.}

\item{model}{vector; model to export driving data. Options include c("GOTM", "GLM", "Simstrat",
"FLake")}

\item{res_files}{a vector with the paths to the parameter and result files generated by
`cali_ensemble()` (the return value of running `cali_ensemble()` with cmethod = "LHC").}
}
\description{
Plot the results of a calibration done using the `cali_ensemble()` function with cmethod = "LHC".
}
