% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_var.R
\name{load_var}
\alias{load_var}
\title{Extract variables from NetCDF file}
\usage{
load_var(ncdf, var, return = "list", print = TRUE)
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{var}{character; Name of the variable to be extracted. Must match short name in netCDF file}

\item{return}{character; Must be either list or array}

\item{print}{logical; Print the name and units of the variable extracted, defaults to TRUE}
}
\value{
dataframe in the same format as the observation file with the surface in the top column and the bottom in the last column.
}
\description{
Extracts a selected parameter from the netCDF file and formats it into a dataframe.
}
