% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calibration.R
\name{make_script}
\alias{make_script}
\title{Creates a script to then run as a job}
\usage{
make_script(call, name)
}
\arguments{
\item{call}{an object of class `call``}

\item{name}{name of the job}
}
\description{
Creates a script to then run as a job
}
\keyword{internal}
