% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_temp_dir.R
\name{make_temp_dir}
\alias{make_temp_dir}
\title{Create temporary directories from model folders.}
\usage{
make_temp_dir(model, folder = ".", n = 2, tmp_dir = NULL)
}
\arguments{
\item{model}{string; for which model. Options are c("dy_cd", "glm_aed" and
"gotm_wet")}

\item{folder}{filepath; to directory which contains the model
configuration.}

\item{n}{integer; number of directories to create.}
}
\value{
vector of temporary directories.
}
\description{
Create temporary directories from model folders.
}
