% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_input.R
\docType{data}
\name{met_input}
\alias{met_input}
\title{Description of meteorological inputs.}
\format{
## `met_input`
A data frame with 19 rows and 4 columns:
\describe{
  \item{Description}{Name of variable}
  \item{Units}{Unit variable needs to be supplied in}
  \item{Column Name}{Column name to be used in LakeEnsemblR input file}
  \item{Status}{Description of if the variable is a required variable or can be calculated internally}
}
}
\source{
Developer generated.
}
\usage{
met_input
}
\description{
Description of meteorological inputs.
}
\keyword{datasets}
