% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_var_dic.R
\docType{data}
\name{met_var_dic}
\alias{met_var_dic}
\title{Meteorological variable dictionary.}
\format{
## `met_var_dic`
A data frame with 21 rows and 2 columns:
\describe{
  \item{standard_name}{Name used for each variable}
  \item{short_name}{Short name used within the code and netCDF file}
}
}
\source{
Developer generated.
}
\usage{
met_var_dic
}
\description{
Meteorological variable dictionary.
}
\keyword{datasets}
