% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_var_dic.R
\docType{data}
\name{model_var_dic}
\alias{model_var_dic}
\title{Model variable dictionary.}
\format{
## `model_var_dic`
A data frame with 21 rows and 2 columns:
\describe{
  \item{model}{Lake model}
  \item{var}{Short variable name used within the lake model}
  \item{longname}{Long variable name used within the lake model}
  \item{units}{Units}
  \item{dims}{Dimensions}
  \item{file}{Filepath to model output}
  \item{Plot}{Logical; should this variable be included in plotting output}
}
}
\source{
Developer generated.
}
\usage{
model_var_dic
}
\description{
Model variable dictionary.
}
\keyword{datasets}
