% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble.R
\name{plot_ensemble}
\alias{plot_ensemble}
\title{Function to plot results of the ensemble run}
\usage{
plot_ensemble(
  ncdf,
  model = c("FLake", "GLM", "GOTM", "Simstrat", "MyLake"),
  var,
  depth = NULL,
  date = NULL,
  av_fun = "mean",
  boxwhisker = FALSE,
  residuals = FALSE
)
}
\arguments{
\item{ncdf}{Path to the netcdf file created by `run_ensemble()`}

\item{model}{Vector of models which should be included in the plot}

\item{var}{Variable which to plot}

\item{depth}{If `var` has a depth dimension, for which depth should it be plotted?}

\item{date}{Specific date for which depth profiles should be plotted}

\item{av_fun}{Averaging function to use, defaults to the arithmetic mean (`mean()`)}
}
\description{
Plot the outcome of the ensemble run for a given depth along with the minimum, maximum, and
an average value.
}
\examples{
\dontrun{
# time series
p1 <- plot_ensemble(ncdf = "output/ensemble_output.nc",
                   model = c('FLake', 'GLM',  'GOTM', 'Simstrat', 'MyLake'),
                   var = "watertemp", depth = 0.9, boxwhisker = TRUE)

# depth profiles
p2 <- plot_ensemble(ncdf = "output/ensemble_output.nc",
                   model = c('FLake', 'GLM',  'GOTM', 'Simstrat', 'MyLake'),
                   var = "watertemp", date = as.POSIXct("2010-06-13", tz = "UTC"),
                   boxwhisker = TRUE)
                   
}

}
\author{
Johannes Fledbauer
}
