% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LHC.R
\name{plot_LHC}
\alias{plot_LHC}
\title{Plot results of LHC calibration}
\usage{
plot_LHC(config_file, model, res_files, qual_met = "rmse", best_quant = 0.1)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file for which `cali_ensemble()`
was run.}

\item{model}{vector; model to export driving data. Options include c("GOTM", "GLM", "Simstrat",
"FLake")}

\item{res_files}{a vector with the paths to the parameter and result files generated by
`cali_ensemble()` (the return value of running `cali_ensemble()` with cmethod = "LHC").}

\item{qual_met}{Name of the performance metric to use. ;ust be a column name in the model
result files.}

\item{best_quant}{Quantile for wich the distribution of the parameetrs is to be plotted}
}
\description{
Plot the results of a calibration done using the `cali_ensemble()` function with cmethod = "LHC".
}
