% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw.R
\name{plot_raw}
\alias{plot_raw}
\title{Plot raw model output}
\usage{
plot_raw(con, model, var, xlim = NULL, zlim = NULL, var_lab = NULL)
}
\arguments{
\item{con}{list; output from `connect_output()`.}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{var}{string; short variable name within the model. Found in the "var" column in `model_var_dic`.}

\item{xlim}{Date; pairs for min and max limits for the x-axis. Default is NULL and will use the range in the dataset.}

\item{zlim}{Date; pairs for min and max limits for the Z-axis. Default is NULL and will use the range in the dataset.}

\item{var_lab}{string; label to be used for y-axis (2-D plot) or z-axis (3-D plot). Default is NULL and will use `var` value.}
}
\value{
list
}
\description{
Plot a variable from the raw model output.
}
\examples{
\dontrun{
}

}
