% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calibration.R
\name{qual_fun}
\alias{qual_fun}
\title{calculate model performance metrics}
\usage{
qual_fun(O, P)
}
\arguments{
\item{O}{data.frame containing observed values, first row is datetime}

\item{P:}{data.frame containing predicted values, first row is datetime}
}
\description{
function that calculates different estimations for model accuracy, namely: root mean squared
error (rmse), (Nash-Sutcliff) model efficiency (nse), Pearson corelation coefficient (r),
relative error (re) and normalized mean absolute error (nmae). returns  a data.frame
containing the five quality estimates
}
\keyword{internal}
