% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{quality_measures}
\alias{quality_measures}
\title{Calculate model performance metrics}
\usage{
quality_measures(ncdf, model, var, qualfun = qual_meas)
}
\arguments{
\item{ncdf}{Path to the netcdf file generated by running `run_ensemble`}

\item{model}{Vector of models for which to calculate the performance measures}

\item{var}{Variable for which to calculate the performance measures}

\item{qualfun}{Function to calculate the performance measures. Per default calculates root
mean suqared error (rmse), Nash-Shutcliff efficiency (nse), Pearson correlation (r),
relative error (re), and normalized mean absolute error (nmae). Can be any function that
takes observed data as first, and simulated data at the same time and depth as the second
argument}
}
\description{
Claculates model performace metrics from the netcdf file generated by running `run_ensemble`
}
\examples{
\dontrun{
quality_measures(ncdf = "output/ensemble_output.nc",
                   model = c("FLake", "GLM",  "GOTM", "Simstrat", "MyLake"),
                   var = "watertemp")
}
}
\author{
Johannes Feldbauer
}
