% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flake_out.R
\name{read_flake_out}
\alias{read_flake_out}
\title{Extract FLake output}
\usage{
read_flake_out(
  output,
  vars,
  depths,
  folder = ".",
  nml_file,
  long = FALSE,
  out_time,
  out_hour = 0
)
}
\arguments{
\item{output}{filepath; to FLake output file}

\item{vars}{vector; variables to extract from FLake output. Currently just temp and ice}

\item{depths}{vector; of numeric values to extract the depths at. Only used if extracting water temp}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{nml_file}{filepath; to FLake namelist file}

\item{long}{Boolean; return data in long form}

\item{out_time}{vector; of output time values to subset data by.}

\item{out_hour}{numeric; hour of output time values to subset data. Only used for FLake if model time step is 86400s.}
}
\value{
Dataframe if only one variable otherwise a list
}
\description{
Read in FLake results, interpolate to a gridded struture and return a dataframe in long or wide format for vars = "temp"
}
\examples{
\dontrun{
df <- read_flake_out(output = "FLake/output/output.dat", folder = "FLake", nml_file = "FLake/feeagh.nml")
}

}
