% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.R
\name{read_restart}
\alias{read_restart}
\title{Read model restart variables}
\usage{
read_restart(folder = ".", model)
}
\arguments{
\item{folder}{folder}

\item{model}{string; denoting which model you are extracting from. Can be 'GLM', 'GOTM', 'Simstrat'}
}
\description{
Read in model restart variables in a list format
}
\examples{


}
