% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yaml.R
\name{read_yaml}
\alias{read_yaml}
\title{Read a YAML file}
\usage{
read_yaml(file, fileEncoding = "UTF-8", text, error.label, ...)
}
\arguments{
\item{file}{filepath; to yaml file which you wish to edit}

\item{fileEncoding}{character string: if non-empty declares the encoding used on a file (not a connection) so the character data can be re-encoded. See file.}

\item{text}{character string: if file is not supplied and this is, then data are read from the value of text via a text connection. Notice that a literal string can be used to include (small) data sets within R code.}

\item{error.label}{a label to prepend to error messages (see Details).}

\item{...}{arguments to pass to yaml.load}
}
\value{
A list with the yaml file
}
\description{
Read a YAML document from a file and create an R object from it. This is from the `yaml` package and is built to replicate the functionality used in the `glmtools`
package.
}
\examples{

\dontrun{
config_file <- system.file("extdata/feeagh/LakeEnsemblR.yaml", package = "LakeEnsemblR")

yaml <- read_yaml(config_file)
yaml <- set_yaml(yaml, value = 23, key1 = "location", key2 = "latitude")
yaml <- set_yaml(yaml, value = "2010-06-01 00:00:00", key1 = "time", key2 = "start")
yaml <- set_yaml(yaml, value = "meteo.csv", key1 = "input", key2 = "meteo", key3 = "file")
yaml <- set_yaml(yaml, value = TRUE, key1 = "calibration", key2 = "GOTM", key3 = "turb_param/k_min", key4 = "log")
yaml <- set_yaml(yaml, value = c("temp", "salt"), key1 = "output", key2 = "variables")

write_yaml(yaml, "LakeEnsemblR.yaml")
}
}
\author{
Jeremy Stephens <jeremy.f.stephens@vumc.org>
}
