% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run Ensemble of lake models}
\usage{
run_ensemble(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake"),
  folder = ".",
  return_list = FALSE,
  create_netcdf = TRUE,
  obs_file = NULL
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{return_list}{boolean; Return a list of dataframes of model output. Deafults to FALSE}

\item{create_netcdf}{boolean; Create ensemble_output.nc4 file}

\item{obs_file}{filepath; to LakeEnsemblR standardised observed water temperature profile data. If included adds observed data to netCDF and list if they are set to TRUE. Defaults to NULL.}
}
\description{
Run each of the lake models
}
