% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LHC.R
\name{run_LHC}
\alias{run_LHC}
\title{Run Latin hypercube sampling}
\usage{
run_LHC(
  config_file,
  num = NULL,
  param_file = NULL,
  method,
  model = c("FLake", "GLM", "GOTM", "Simstrat"),
  folder = ".",
  spin_up = NULL
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{num}{integer; the number of random parameter sets to generate. If param file is provided
num = number of parameters in that file.}

\item{param_file}{filepath; to previously created parameter file set. If NULL creates a new
parameter set. Defaults to NULL
#' @param method character; Method for calibration. Can be "met", "model" or "both". Needs to be
specified by the user.}

\item{model}{vector; model to export driving data. Options include c("GOTM", "GLM", "Simstrat",
"FLake")}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{spin_up}{numeric; Number of days to disregard as spin-up for analysis.}

\item{obs_file}{filepath; to LakeEnsemblR standardised observed water temperature profile data.
If included adds observed data to netCDF and list if they are set to TRUE. Defaults to NULL.}

\item{meteo_file}{filepath; to met file which is in the standardised LakeEnsemblR format.}
}
\description{
Run lake models using Latin hypercube sampling for model parameters.
}
\examples{
\dontrun{
pars <- c("wind_factor", "swr_factor", "lw_factor")
mat <- matrix(data = c(0.5,2,0.5,1.5,0.5,1.5), nrow = 3, byrow = T)
df <- as.data.frame(mat)
rownames(df) <- pars
run_LHC(num = 100, obs_file = "LakeEnsemblR_wtemp_profile_standard.csv",
config_file = "Feeagh_master_config.yaml", model = "FLake", meteo_file =
"LakeEnsemblR_meteo_standard.csv")
pars <- c("wind_factor", "swr_factor", "lw_factor")
mat <- matrix(data = c(0.5,2,0.5,1.5,0.5,1.5), nrow = 3, byrow = TRUE)
df <- as.data.frame(mat)
rownames(df) <- pars
run_LHC(num = 100, obs_file = "LakeEnsemblR_wtemp_profile_standard.csv",
config_file = "Feeagh_master_config.yaml", model = "FLake", meteo_file =
"LakeEnsemblR_meteo_standard.csv")
}
}
