% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LHC_atom.R
\name{run_LHC_atom}
\alias{run_LHC_atom}
\title{Run Latin hypercube sampling atomized version}
\usage{
run_LHC_atom(
  config_file,
  num = NULL,
  param_file = NULL,
  method,
  qualfun = qual_meas,
  model = c("FLake", "GLM", "GOTM", "Simstrat"),
  folder = ".",
  spin_up = NULL,
  out_f = "cali",
  nout_fun = 5
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{num}{integer; the number of random parameter sets to generate. If param file is provided
num = number of parameters in that file.}

\item{param_file}{filepath; to previously created parameter file set. If NULL creates a new
parameter set. Defaults to NULL}

\item{method}{character; Method for calibration. Can be "met", "model" or "both". Needs to be
specified by the user.}

\item{qualfun}{function; function that calculates measure of fit from observed and simulated 
variables, takes the two arguments Observed and Simulated}

\item{model}{vector; model to export driving data. Options include c("GOTM", "GLM", "Simstrat",
"FLake")}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{spin_up}{numeric; Number of days to disregard as spin-up for analysis.}

\item{out_f}{character; name of the folder to store results into}

\item{nout_fun}{integer; number of return values from qualfun}

\item{obs_file}{filepath; to LakeEnsemblR standardised observed water temperature profile data.
If included adds observed data to netCDF and list if they are set to TRUE. Defaults to NULL.}
}
\description{
Run lake models using Latin hypercube sampling for model parameters.
}
\examples{
\dontrun{

config_file <- 'LakeEnsemblR.yaml'

run_LHC_atom(config_file = config_file, num = 200, method = "met",
             model = c("FLake", "GLM", "GOTM", "Simstrat", "MyLake"))
}
}
