% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LHC_parallel.R
\name{run_LHC_parallel}
\alias{run_LHC_parallel}
\title{Run Latin hypercube sampling in parallel}
\usage{
run_LHC_parallel(
  config_file,
  num = NULL,
  param_file = NULL,
  method,
  model = c("FLake", "GLM", "GOTM", "Simstrat"),
  folder = ".",
  spin_up = NULL,
  num_cores = NULL
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{num}{integer; the number of random parameter sets to generate. If param file is provided
num = number of parameters in that file.}

\item{param_file}{filepath; to previously created parameter file set. If NULL creates a new
parameter set. Defaults to NULL}

\item{method}{character; Method for calibration. Can be "met", "model" or "both". Needs to be
specified by the user.}

\item{model}{vector; model to export driving data. Options include c("GOTM", "GLM", "Simstrat",
"FLake")}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{spin_up}{numeric; Number of days to disregard as spin-up for analysis.}

\item{obs_file}{filepath; to LakeEnsemblR standardised observed water temperature profile data.
If included adds observed data to netCDF and list if they are set to TRUE. Defaults to NULL.}

\item{meteo_file}{filepath; to met file which is in the standardised LakeEnsemblR format.}
}
\description{
Call run_LHC with parallel::clusterApply
}
