% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lint.R
\name{run_lint}
\alias{run_lint}
\title{Runs lint for LER script}
\usage{
run_lint(file)
}
\arguments{
\item{file}{filepath; to R script that you want to run the lintr::lint function on}
}
\description{
Runs lintr::lint function, using the settings agreed upon for LakeEnsemblR
}
\examples{
\dontrun{
run_lint(file = "export_config.R")
}
}
