% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCMC.R
\name{run_MCMC}
\alias{run_MCMC}
\title{Run pseudo-MCMC sampling}
\usage{
run_MCMC(
  config_file,
  num = 100,
  param_file = NULL,
  method,
  model = c("FLake", "GLM", "GOTM", "Simstrat"),
  folder = ".",
  spin_up = NULL
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{num}{integer; the number of iterations the algorithm should run}

\item{param_file}{filepath; defaults to NULL}

\item{method}{character; Method for calibration. Can be "met", "model" or "both". Needs to be
specified by the user.}

\item{model}{vector; model to export driving data. Options include c("GOTM", "GLM", "Simstrat",
"FLake")}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{spin_up}{numeric; Number of days to disregard as spin-up for analysis.}

\item{obs_file}{filepath; to LakeEnsemblR standardised observed water temperature profile data.
If included adds observed data to netCDF and list if they are set to TRUE. Defaults to NULL.}

\item{meteo_file}{filepath; to met file which is in the standardised LakeEnsemblR format.}
}
\description{
Run lake models using a pseudo Markov Chain Monte Carlo algorithm.
}
