% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_yaml.R
\name{set_yaml}
\alias{set_yaml}
\title{Set a value within the loaded yaml file}
\usage{
set_yaml(yaml, value, ...)
}
\arguments{
\item{yaml}{list; loaded using `read_yaml()`}

\item{value}{string; to be input into the the yaml file.
Note boolean values must be input as "true"/"false" as per the json format}

\item{...}{string key1, key2, etc.: multiple keys pointing toward the line
that you want to edit in the yaml file. Keys must be listed consecutively,
without skipping numbers.}
}
\value{
list with updated yaml parameters
}
\description{
Set the value within a yaml file.
}
\examples{

\dontrun{
config_file <- system.file("extdata/feeagh/LakeEnsemblR.yaml", package = "LakeEnsemblR")

yaml <- read_yaml(config_file)
yaml <- set_yaml(yaml, value = 23, key1 = "location", key2 = "latitude")
yaml <- set_yaml(yaml, value = "2010-06-01 00:00:00", key1 = "time", key2 = "start")
yaml <- set_yaml(yaml, value = "meteo.csv", key1 = "input", key2 = "meteo", key3 = "file")
yaml <- set_yaml(yaml, value = TRUE, key1 = "calibration", key2 = "GOTM", key3 = "turb_param/k_min", key4 = "log")

write_yaml(yaml, "LakeEnsemblR.yaml")
}
}
\author{
Tadhg Moore
}
