% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_average.R
\name{time_average}
\alias{time_average}
\title{Calculates the average of an irregular time series}
\usage{
time_average(timeseries, start, end, n = 1000)
}
\arguments{
\item{timeseries}{dataframe; first column POSIXct, second column numeric}

\item{start}{POSIXct; start of the averaging window}

\item{end}{POSIXct; end of the averaging window}

\item{n}{integer; number of blocks}
}
\description{
Calculates the average of an irregular time series
}
\keyword{methods}
