% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calibration.R
\name{wrap_model}
\alias{wrap_model}
\title{warpper function for other two methods (modMCMC and modFit)}
\usage{
wrap_model(
  pars,
  type,
  model,
  var,
  config_file,
  met,
  folder,
  out_f,
  obs_deps,
  obs_out,
  out_hour,
  qualfun,
  config_f,
  outf_n,
  write = TRUE
)
}
\arguments{
\item{pars}{named vector with a set of parameter for wich the model is to be run and a
performance metric is to be returned}

\item{type}{character vector specifying the type of parameter. can be either `"met"` for meteo
scaling or "model" for model specific parameters, must have the same length as `pars`.}

\item{model}{the model name for which to run the function}

\item{var}{variable for which to calculate the model performance (usually this is `"temp"`)}

\item{config_file}{path to the master config file}

\item{met}{data.frame with the model specific meteo data, which is then feed to scale_meteo}

\item{folder}{root folder}

\item{out_f}{folder in the root folder where the output of the LHC should be written to}

\item{obs_deps}{depths of observations}

\item{obs_out}{data.frame with the observations of `var`.}

\item{out_hour}{FLake specific: hout of output}

\item{qualfun}{function that takes data.frames of observations and simulations and returns
model performance metrics}

\item{outf_n}{name of the output file in which the model performance for every parameter
set is written}

\item{write}{write the results to a file?}

\item{nout_fun}{number of values returned by `qualfun`.}
}
\description{
A wrapper function for the modMCMC and modFit calibration methods, used in cali_ensemble when
cmethod == "modMCMC" or "modFit".
}
\keyword{internal}
