% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.R
\name{write_restart}
\alias{write_restart}
\title{Write model restart variables}
\usage{
write_restart(folder = ".", model, restart_list)
}
\arguments{
\item{folder}{folder}

\item{model}{string; denoting which model you are extracting from. Can be 'GLM', 'GOTM', 'Simstrat'}
}
\description{
Write model restart variables using the list from `read_restart()`
}
