% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_strat.R
\name{analyse_strat}
\alias{analyse_strat}
\alias{analyze_strat}
\title{Returns stratification statstics}
\usage{
analyse_strat(data = NULL, Ts, Tb, dates, H_ice = NULL, drho = 0.1, NH = TRUE)

analyze_strat(data = NULL, Ts, Tb, dates, H_ice = NULL, drho = 0.1, NH = TRUE)
}
\arguments{
\item{data}{dataframe; water temperature data in long format with date, depths, value. Defaults
to NULL}

\item{Ts}{vector;  of surface temperatures which corresponds to date vector}

\item{Tb}{vector; of bottom temperatures which corresponds to date vector}

\item{dates}{vector; of POSIX style date corresponding to rows of Ts and Tb.}

\item{H_ice}{vector; of ice thickness which corresponds to date vector, set to NULL if analysis
not required. Defaults to NULL}

\item{drho}{numeric; density difference between top and bottom indicating stratification
[kg m^-3]}

\item{NH}{boolean; northern hemisphere? TRUE or FALSE. Defaults to true}
}
\description{
Returns stratification statstics: annual mean, max and total
length of summer, winter stratification and ice duration.
NOTE: summer strat periods are allocated to the year in which the period starts. Winter
stratification and ice periods are allocated to the year in which they end.
}
\examples{
\dontrun{
strat <- analyse_strat(Ts = df[,2], Tb = df[,ncol(df)], dates = df[,1])
}

}
\author{
Tom Shatwell
}
