% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fit.R
\name{calc_fit}
\alias{calc_fit}
\title{Calculate model performance metrics}
\usage{
calc_fit(ncdf, list = NULL, model, var, qualfun = qual_meas, avfun = "mean")
}
\arguments{
\item{ncdf}{Path to the netcdf file generated by running `run_ensemble`}

\item{list}{Alternatively to `ncdf` a list of siimulated variables, as returned by
`run_ensemble()` when argument `return_list = TRUE`}

\item{model}{Vector of models for which to calculate the performance measures}

\item{var}{Variable for which to calculate the performance measures}

\item{qualfun}{Function to calculate the performance measures. Per default calculates root
mean suqared error (rmse), Nash-Shutcliff efficiency (nse), Pearson correlation (r),
relative error (re), and normalized mean absolute error (nmae). Can be any function that
takes observed data as first, and simulated data at the same time and depth as the second
argument}

\item{avfun}{Name of the function to calculate the ensemble average, defaults to "mean"}
}
\description{
Calculates model performace metrics from the netcdf file generated by running `run_ensemble`
}
\examples{
\dontrun{
# using standard quality measures
calc_fit(ncdf = "output/ensemble_output.nc",
                   model = c("FLake", "GLM",  "GOTM", "Simstrat", "MyLake"),
                   var = "watertemp")
# using own performance measure
 calc_fit(ncdf = "output/ensemble_output.nc",
                  model = c("FLake", "GLM",  "GOTM", "Simstrat", "MyLake"),
                  var = "watertemp", qualfun = function(O, S) mean(O - S, na.rm = TRUE))
}
}
\author{
Johannes Feldbauer
}
