% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali_ensemble.R
\name{cali_ensemble}
\alias{cali_ensemble}
\title{Calibrate models}
\usage{
cali_ensemble(
  config_file,
  num = NULL,
  param_file = NULL,
  cmethod = "LHC",
  qualfun = qual_fun,
  model = c("FLake", "GLM", "GOTM", "Simstrat", "MyLake"),
  folder = ".",
  spin_up = NULL,
  out_f = "cali",
  nout_fun = 5,
  ...
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{num}{integer; the number of random parameter sets to generate. If param file is provided
num = number of parameters in that file.}

\item{param_file}{filepath; to previously created parameter file set. If NULL creates a new
parameter set. Defaults to NULL}

\item{cmethod}{character; Method for calibration. Can be "LHC", "MCMC" or "". Defaults to "LHC"}

\item{qualfun}{function; function that calculates measure of fit from observed and simulated 
variables, takes the two arguments Observed and Simulated}

\item{model}{vector; model to export driving data. Options include c("GOTM", "GLM", "Simstrat",
"FLake")}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{spin_up}{numeric; Number of days to disregard as spin-up for analysis.}

\item{out_f}{character; name of the folder to store results into}

\item{nout_fun}{integer; number of return values from qualfun}

\item{...}{additional arguments passed to modFit or modMCMC. Only used when method is
modFit or MCMC}

\item{obs_file}{filepath; to LakeEnsemblR standardised observed water temperature profile data.
If included adds observed data to netCDF and list if they are set to TRUE. Defaults to NULL.}
}
\description{
Use one of three methods to calibrate specified models.
}
\examples{
\dontrun{

config_file <- 'LakeEnsemblR.yaml'

cali_ensemble(config_file = config_file, num = 200, cmethod = "LCH",
             model = c("FLake", "GLM", "GOTM", "Simstrat", "MyLake"))
             
resMCMC <- cali_ensemble(config_file = config_file, num = 200, cmethod = "MCMC",
                         model = c("FLake", "GLM", "GOTM", "Simstrat", "MyLake"))
               
resMmodFit <- cali_ensemble(config_file = config_file, num = 200, cmethod = "modFit",
                            model = c("FLake", "GLM", "GOTM", "Simstrat", "MyLake"),
                            method = "Nelder-Mead")                                           
             
}
}
