% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calibration.R
\name{change_pars}
\alias{change_pars}
\title{Change parameter or meteo scaling for a model}
\usage{
change_pars(config_file, model, pars, type, met, folder)
}
\arguments{
\item{config_file}{path to master config file}

\item{model}{name of the model}

\item{pars}{named vector of parameters to change}

\item{type}{character vector specifying the type of parameter. can be either `"met"` for meteo
scaling or "model" for model specific parameters, must have the same length as `pars`}

\item{met}{data.frame with the model specific meteo data, which is then feed to scale_meteo}

\item{folder}{root folder}
}
\description{
Input a specific parameter or scale the meteorological forcing for a selected model
}
\keyword{internal}
