% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{check_master_config}
\alias{check_master_config}
\title{Check master config file}
\usage{
check_master_config(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  exp_cnf = FALSE
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{exp_cnf}{boolean; check if the control files for the models are there}
}
\description{
Check if the master config file is correct
}
