% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali_ensemble.R
\name{cost_model}
\alias{cost_model}
\title{Run a model and calculate model fit}
\usage{
cost_model(
  config_file,
  model,
  var,
  folder,
  obs_deps,
  obs_out,
  out_hour,
  qualfun,
  config_f
)
}
\description{
Runns the selected model and calculates fit metrics using a provided funtion
}
