% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_config.R
\name{export_config}
\alias{export_config}
\title{Export model configuration setups}
\usage{
export_config(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  meteo = TRUE,
  init_cond = TRUE,
  light_ext = TRUE,
  model_param = TRUE,
  folder = ".",
  inflow_file = NULL
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file.
Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{meteo}{boolean; export meteorology data. Defaults to TRUE.}

\item{init_cond}{boolean; export initial conditions. Defaults to TRUE.}

\item{light_ext}{boolean; export light extinction data. Defaults to TRUE.}

\item{model_param}{boolean; export model parameters specificed in the yaml
configuration file. Defaults to TRUE.}

\item{folder}{folder}

\item{inflow_file}{filepath; to inflow file which is in the standardised LakeEnsemblR format (if
a different file than the one provided in the configuration file is needed); default is NULL}
}
\description{
Create directory with file setups for each model based on a master LakeEnsemblR config file
}
\examples{


}
\author{
Tadhg Moore, Jorrit Mesman, Johannes Feldbauer, Robert Ladwig
}
\keyword{methods}
