% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_config.R
\name{export_config}
\alias{export_config}
\title{Export model configuration setups}
\usage{
export_config(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  dirs = TRUE,
  time = TRUE,
  location = TRUE,
  output_settings = TRUE,
  meteo = TRUE,
  init_cond = TRUE,
  extinction = TRUE,
  flow = TRUE,
  model_parameters = TRUE,
  folder = "."
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{dirs}{boolean; create directories for each model and if needed copies templates.
Calls export_dirs. Defaults to TRUE}

\item{time}{boolean; exports time settings. Calls export_time. Defaults to TRUE.}

\item{location}{boolean; exports location and hypsograph settings.
Calls export_location. Defaults to TRUE.}

\item{output_settings}{boolean; exports output settings.
Calls export_output_settings. Defaults to TRUE.}

\item{meteo}{boolean; export meteorology data.
Calls export_meteo. Defaults to TRUE.}

\item{init_cond}{boolean; export initial conditions.
Calls export_init_cond. Defaults to TRUE.}

\item{extinction}{boolean; export light extinction data.
Calls export_extinction. Defaults to TRUE.}

\item{flow}{boolean; export flow settings.
Calls export_flow. Defaults to TRUE.}

\item{model_parameters}{boolean; export model parameters specificed in the yaml
configuration file. Calls export_model_parameters. Defaults to TRUE.}

\item{folder}{folder}
}
\description{
Exports all settings from the master LakeEnsemblR config file
}
\examples{
\dontrun{
config_file <- "LakeEnsemblR.yaml"
export_config(config_file = config_file,
              model = c("FLake", "GLM", "GOTM", "Simstrat", "MyLake"),
              folder = ".")
}

}
\author{
Tadhg Moore, Jorrit Mesman, Johannes Feldbauer, Robert Ladwig
}
\keyword{methods}
