% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dirs.R
\name{export_dirs}
\alias{export_dirs}
\title{Creates directories for each model}
\usage{
export_dirs(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}
}
\description{
Creates directories with file setups for each model, based on the master LakeEnsemblR config file
}
\keyword{methods}
