% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_extinction.R
\name{export_extinction}
\alias{export_extinction}
\title{Export extinction coefficients}
\usage{
export_extinction(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = "."
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{folder}
}
\description{
Exports extinction coefficients for each model based on a master LakeEnsemblR config file
}
\examples{


}
\keyword{methods}
