% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_inflow.R
\name{export_inflow}
\alias{export_inflow}
\title{Export LakeEnsemblR standardised inflow files to model specific driver format}
\usage{
export_inflow(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = "."
)
}
\arguments{
\item{config_file}{name of the master LakeEnsemblR config file}

\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{filepath; to folder which contains the model folders generated
by export_config()}
}
\description{
Export inflow driver files for each model
}
\examples{
\dontrun{
export_inflow(config_file, model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"))
}
}
