% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_inflow.R
\name{export_inflow}
\alias{export_inflow}
\title{Export LakeEnsemblR standardized input to model specific driver format}
\usage{
export_inflow(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  scale_param = NULL,
  folder = ".",
  use_outflows = FALSE
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include
c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{scale_param}{dataframe; with column names referring to standard names in LER.
Scales meteorological value according to value in the column. Defaults to NULL.}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}
}
\description{
Export inflow driver files for each model
}
\examples{
\dontrun{
export_meteo(model = c('GOTM', 'GLM', 'Simstrat', 'FLake'),
             meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
}
