% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_init_cond.R
\name{export_init_cond}
\alias{export_init_cond}
\title{Export initial conditions into each model setup from the LakeEnsemblR standardized input for observed temperature profile}
\usage{
export_init_cond(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake"),
  wtemp_file = NULL,
  date,
  month = NULL,
  ndeps = 2,
  btm_depth = NULL,
  print = TRUE,
  folder = "."
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{wtemp_file}{filepath; to met file which is in the standardised LakeEnsemblR format. If NULL it uses the file from config_file. Defaults to NULL.}

\item{date}{character; Date in "YYYY-mm-dd HH:MM:SS" format to extract the initial profile. If using month, the date to which to set the start date}

\item{month}{numeric;select a month if you want to use an 'average profile from a particular month. Defaults to NULL.}

\item{ndeps}{numeric; number of depths to extract from the monthly average profile. Defaults to 2 (surface and bottom)}

\item{btm_depth}{numeric; Depth to extract the bottom temperature from, must be negative. If none provided uses the max depth in the observed file. Defaults to NULL}

\item{print}{logical; Prints the temperature profile to the console}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}
}
\description{
Export initial condition files and input into model configuration files.
}
