% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_meteo.R
\name{export_meteo}
\alias{export_meteo}
\title{Export LakeEnsemblR standardized input to model specific driver format}
\usage{
export_meteo(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = "."
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include
c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}
}
\description{
Export driver files for each model
}
\examples{
\dontrun{
export_meteo(model = c('GOTM', 'GLM', 'Simstrat', 'FLake'),
             meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
}
