% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_meteo.R
\name{export_meteo}
\alias{export_meteo}
\title{Export LakeEnsemblR standardized input to model specific driver format}
\usage{
export_meteo(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake"),
  meteo_file,
  lhc_file = NULL,
  metric = "RMSE",
  folder = "."
)
}
\arguments{
\item{model}{vector; model to export driving data. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{meteo_file}{filepath; to met file which is in the standardised LakeEnsemblR format.}

\item{lhc_file}{filepath to results from Latin Hypercube sampling from run_Latin_hypercube(). Defaults to NULL}

\item{metric}{character; metric to extract best scaling factors, can be 'RMSE', 'NSE', 'Pearson_r', 'MAE'. Defaults to RMSE}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}
}
\description{
Export driver files for each model
}
\examples{
\dontrun{
export_meteo(model = c('GOTM', 'GLM', 'Simstrat', 'FLake'), meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
}
