% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_meteo.R
\name{export_meteo}
\alias{export_meteo}
\title{Export LakeEnsemblR standardized input to model specific driver format}
\usage{
export_meteo(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake"),
  meteo_file,
  folder = "."
)
}
\arguments{
\item{model}{vector; model to export driving data. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{meteo_file}{filepath; to met file which is in the standardised LakeEnsemblR format.}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}
}
\description{
Export driver files for each model
}
\examples{
\dontrun{
export_meteo(model = c('GOTM', 'GLM', 'Simstrat', 'FLake'), meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
}
