% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Get output data for each model}
\usage{
get_output(
  config_yaml,
  model,
  vars,
  obs_depths = NULL,
  folder = ".",
  out_time,
  out_hour,
  run_success
)
}
\arguments{
\item{config_yaml}{list; loaded using `read_yaml()`}

\item{model}{character; Model for which scaling parameters will be applied. Options include
c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{vars}{vector; variables to extract from FLake output. Currently just temp and ice}

\item{obs_depths}{vector; Observation depths. Defaults to NULL}

\item{folder}{filepath; to folder which contains the model folders generated by export_config().
Defaults to '.'}

\item{out_time}{vector; of output time values to subset data by.}

\item{out_hour}{numeric; hour of output time values to subset data. Only used for FLake if model
time step is 86400s.}

\item{run_success}{logical; was the model run successful?}
}
\value{
dataframe or list of output variables
}
\description{
Get output data for each model
}
