% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_config_value.R
\name{input_config_value}
\alias{input_config_value}
\title{Inputs value into any LER model config file}
\usage{
input_config_value(model, yaml)
}
\arguments{
\item{model}{vector; model to export configuration file for.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{yaml}{list; loaded using `read_yaml()`}
}
\description{
Wrapper function for functions that input values into model config files
}
\examples{
\dontrun{
input_config_value(model = "GOTM", file = "gotm.yaml",
  label = "turb_param", key = "k_min", value = 0.00001)
}

}
