% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_config_value.R
\name{input_config_value}
\alias{input_config_value}
\title{Inputs value into any LER model config file}
\usage{
input_config_value(model, file, label = NULL, key = NULL, value, ...)
}
\arguments{
\item{model}{string; name of a model in the LER package}

\item{file}{filepath; to model-specific config file}

\item{label}{string; which corresponds to section where the key is located}

\item{key}{string; name of key in which to extract the value}

\item{value}{string; name of value to input into config file}

\item{...}{string; all required yaml labels
defaults to overwriting file if not specified}
}
\description{
Wrapper function for functions that input values into model config files
}
\examples{
\dontrun{
input_config_value(model = "GOTM", file = "gotm.yaml",
  label = "turb_param", key = "k_min", value = 0.00001)
}

}
