% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heat map of ensemble model output}
\usage{
plot_heatmap(
  ncdf = NULL,
  var = "watertemp",
  dim = "model",
  dim_index = 1,
  var_list = NULL,
  model = NULL
)
}
\arguments{
\item{ncdf}{Path to the netCDF file created by `run_ensemble()`}

\item{var}{Variable which to plot. Defaults to "watertemp"}

\item{dim}{character; NetCDF dimensions to extract. Must be either "member" or "model". Defaults to "model". Only used if plotting from netCDF file. Currently only works with "model".}

\item{dim_index}{numeric; Index of dimension chosen to extract from. Defaults to 1. Only used if plotting from netCDF file.}

\item{var_list}{list of variables in the format when loaded using `load_var()`. Defaults to NULL}

\item{model}{Vector of models which should be included in the plot}
}
\value{
ggplot object of heatmaps
}
\description{
Plot a heat map of ensemble output data. It can either plot directly from the netCDF file or a list in the format when loaded in with `load_var()`.
}
\examples{
\dontrun{
ncdf <- 'output/ensemble_output.nc'
# Plot ensemble mean at 0.9m
plot_ensemble(ncdf = ncdf, model = model, var = 'watertemp', depth = 0.9)
}
}
\author{
Tadhg Moore, Johannes Feldbauer
}
