% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resid.R
\name{plot_resid}
\alias{plot_resid}
\title{Function to plot residuals for each model of the ensemble run}
\usage{
plot_resid(ncdf = NULL, var = "watertemp", var_list = NULL, model = NULL)
}
\arguments{
\item{ncdf}{filepath; to the netcdf file created by `run_ensemble()`}

\item{var}{string; of variable which to plot. Defaults to "watertemp"}

\item{var_list}{list; of variables in the format when loaded using `load_var()`. Defaults to NULL}

\item{model}{string vector; of models which should be included in the plot. If NULL all models in the netCDF/list are plotted. Defaults to NULL.}
}
\value{
list with four ggplot objects: "obs_res" = Observations versus residuals,
 "res_depth" = Residuals versus depth,
 "yday_res" = residuals for day of year,
 "res_dist" = distribution of residuals
}
\description{
Plot residual diagnostic plots. Residuals are calculated by (sim - obs)
 for each corresponding depth and time step there is an observed value.
}
\examples{
\dontrun{
plist <- plot_resid(ncdf = "output/ensemble_output.nc",var = "watertemp",
                   model = c('FLake', 'GLM',  'GOTM', 'Simstrat', 'MyLake'))
plist[['obs_res']]+
theme_classic()
}


}
\author{
Tadhg Moore, Johannes Feldbauer
}
