% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run Ensemble of lake models}
\usage{
run_ensemble(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = ".",
  return_list = FALSE,
  create_netcdf = TRUE
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{return_list}{boolean; Return a list of dataframes of model output. Defaults to FALSE}

\item{create_netcdf}{boolean; Create ensemble output file. Defaults to TRUE}
}
\description{
Run each of the lake models
}
