% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run Ensemble of lake models}
\usage{
run_ensemble(
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
  folder = ".",
  verbose = FALSE,
  parallel = FALSE,
  return_list = FALSE,
  create_netcdf = TRUE,
  add = FALSE
)
}
\arguments{
\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include
c('GOTM', 'GLM', 'Simstrat', 'FLake', 'MyLake')}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{verbose}{Boolean; Should model output be shown in the console. Defaults to FALSE}

\item{parallel}{Boolean; should the model calibration be parallelized}

\item{return_list}{boolean; Return a list of dataframes of model output. Defaults to FALSE}

\item{create_netcdf}{boolean; Create ensemble output file. Defaults to TRUE}

\item{add}{boolean; Add results to an existing netcdf file with new dimension "member"
#' @importFrom parallel detectCores parLapply clusterExport makeCluster stopCluster clusterEvalQ}
}
\description{
Run each of the lake models
}
