% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LHC.R
\name{run_LHC}
\alias{run_LHC}
\title{Run Latin hypercube samplinh}
\usage{
run_LHC(
  parRange,
  num,
  param_file = NULL,
  obs_file,
  config_file,
  model = c("GOTM", "GLM", "Simstrat", "FLake"),
  meteo_file,
  folder = "."
)
}
\arguments{
\item{parRange}{dataframe; the range (min, max) of the parameters, a data.frame with one row for each parameter, and two columns with the minimum (1st) and maximum (2nd) column.}

\item{num}{integer; the number of random parameter sets to generate.}

\item{param_file}{filepath; to previously created parameter file set. If NULL creates a new parameter set. Defaults to NULL}

\item{obs_file}{filepath; to LakeEnsemblR standardised observed water temperature profile data. If included adds observed data to netCDF and list if they are set to TRUE. Defaults to NULL.}

\item{config_file}{filepath; to LakeEnsemblr yaml master config file}

\item{model}{vector; model to export driving data. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{meteo_file}{filepath; to met file which is in the standardised LakeEnsemblR format.}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}
}
\description{
Run lake models using Latin hypercube sampling for model parameters.
}
\examples{
\dontrun{
pars <- c('wind_factor', 'swr_factor', 'lw_factor')
mat <- matrix(data = c(0.5,2,0.5,1.5,0.5,1.5), nrow = 3, byrow = T)
df <- as.data.frame(mat)
rownames(df) <- pars
run_LHC(parRange = parRange, num = 100, obs_file = 'LakeEnsemblR_wtemp_profile_standard.csv', config_file = 'Feeagh_master_config.yaml', model = 'FLake', meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
pars <- c('wind_factor', 'swr_factor', 'lw_factor')
mat <- matrix(data = c(0.5,2,0.5,1.5,0.5,1.5), nrow = 3, byrow = T)
df <- as.data.frame(mat)
rownames(df) <- pars
run_LHC(parRange = parRange, num = 100, obs_file = 'LakeEnsemblR_wtemp_profile_standard.csv', config_file = 'Feeagh_master_config.yaml', model = 'FLake', meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
