% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCMC.R
\name{run_MCMC}
\alias{run_MCMC}
\title{Run pseudo-MCMC sampling}
\usage{
run_MCMC(
  config_file,
  num = 100,
  param_file = NULL,
  method,
  model = c("FLake", "GLM", "GOTM", "Simstrat"),
  folder = ".",
  spin_up = NULL
)
}
\description{
Run lake models using a pseudo Markov Chain Monte Carlo algorithm.
}
