% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_LHC.R
\name{sample_LHC}
\alias{sample_LHC}
\title{Create Latin hypercube sample of parameters}
\usage{
sample_LHC(parRange, num, folder = ".")
}
\arguments{
\item{parRange}{dataframe; the range (min, max) of the parameters, a data.frame with one row for each parameter, and two columns with the minimum (1st) and maximum (2nd) column.}

\item{num}{integer; the number of random parameter sets to generate.}
}
\description{
Create CSV file with parameter samples for the LHC runs
}
\examples{
\dontrun{
pars <- c('wind_factor', 'swr_factor', 'lw_factor')
mat <- matrix(data = c(0.5,2,0.5,1.5,0.5,1.5), nrow = 3, byrow = T)
df <- as.data.frame(mat)
rownames(df) <- pars
param_file <- sample_LHC(parRange = df, num = 300)
run_Latin_hypercube(param_file = param_file, obs_file = 'LakeEnsemblR_wtemp_profile_standard.csv', config_file = 'Feeagh_master_config.yaml', model = 'FLake', meteo_file = 'LakeEnsemblR_meteo_standard.csv')
}
}
