% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_met.R, R/scale_met.R
\name{scale_met}
\alias{scale_met}
\alias{format_met}
\title{Format meteorological data for each model}
\usage{
format_met(met, model, daily = FALSE, config_file, folder = ".")

scale_met(met, pars, model, out_file = NULL)
}
\arguments{
\item{met}{dataframe; in the format for that particular model with the column headers}

\item{model}{character; Model for which scaling parameters will be applied. Options include c('GOTM', 'GLM', 'Simstrat', 'FLake')}

\item{daily}{logical; Is the data on a daily timestep. Defaults to FALSE}

\item{config_file}{filepath; To LER config yaml file. Only used if model = 'GOTM'}

\item{folder}{filepath; to folder which contains the model folders generated by export_config()}

\item{pars}{dataframe; scaling factors for meteorological data with column names corresponding to variable dictionary}

\item{out_file}{filepath; For scaled meteorlogical data}
}
\value{
dataframe of met data in the model format
}
\description{
Format dataframe into shape for the specified model

Scale meteorological parameters.
}
